// src/interfaces/inewcustomer.ts
export interface INewCustomer{
  username : string;
  firstname :string;
  lastname : string;
  dob : string;
  name : string;
  email : string;
  address: string;
  country: string;
  state: string;
  city: string;
  membership: string;
  sex: string;
  phone : string;
  password : string;
  agreement_checkbox: string;
  interests: string[];
}